move operator_client.go to internal
