+++ cmd/bank-vaults/init.go
@@ -18,6 +18,7 @@ import (
 	"github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
 
+	internalVault "github.com/banzaicloud/bank-vaults/internal/vault"
 	"github.com/banzaicloud/bank-vaults/pkg/sdk/vault"
 )
 
@@ -53,7 +54,7 @@ It will not unseal the Vault instance after initialising.`,
 			logrus.Fatalf("error building vault config: %s", err.Error())
 		}
 
-		v, err := vault.New(store, cl, vaultConfig)
+		v, err := internalVault.New(store, cl, vaultConfig)
 		if err != nil {
 			logrus.Fatalf("error creating vault helper: %s", err.Error())
 		}
