+++ cmd/bank-vaults/metrics.go
@@ -15,12 +15,11 @@
 package main
 
 import (
+	internalVault "github.com/banzaicloud/bank-vaults/internal/vault"
 	"github.com/gin-gonic/gin"
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/prometheus/client_golang/prometheus/promhttp"
 	"github.com/sirupsen/logrus"
-
-	"github.com/banzaicloud/bank-vaults/pkg/sdk/vault"
 )
 
 const prometheusNS = "vault"
@@ -56,7 +55,7 @@ var (
 )
 
 type prometheusExporter struct {
-	Vault vault.Vault
+	Vault internalVault.Vault
 	Mode  string
 }
